refactor(api): extract shared ForgotPassword payloads into controllers/shared/auth_entities.py
refactor(api): extract LoginPayloadBase into shared auth entities

refactor(api): move shared auth entities to services/entities/auth_entities.py

fix
